Added "MessingAroundWithLoadTable" file
Despite it's name, this file does not use the loadTable method. At one point it did, but now it uses a BufferedReader to go through the file and make an rray containing trings for each of the points we need to render. Next goal is to get this all working in a Java Program