Base Upload
This commit is done just to have all the files in one place so that I can download them at once on any computer. The .java files can be edited in VSC (where they can be ran through error checking) and the .pde files can be edited in Processing (where they can be run). My assumption now is that when the .java files are functional, they can be put into Processing and run alongside the mapping.